moved the seeds folder into the data folder and updated the knexfile.
