+++ knexfile.js
@@ -7,6 +7,9 @@ module.exports = {
     },
     migrations: {
       directory: "./data/migrations"
+    },
+    seeds: {
+      directory: "./data/seeds"
     }
   }
 };
